fix seed displayer, refacto some components and introduced icons
