+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@headlessui/react": "^2.2.1",
+    "@heroicons/react": "^2.2.0",
     "@tailwindcss/vite": "^4.1.3",
     "@types/bs58": "^4.0.4",
     "@types/uuid": "^10.0.0",
