+++ src/types/transactions.ts
@@ -1,0 +1,1 @@
+export type TransactionId = string;
